--- a/src/libraries/System.Private.CoreLib/src/System/Globalization/DateTimeFormatInfo.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Globalization/DateTimeFormatInfo.cs
@@ -150,7 +150,7 @@ namespace System.Globalization
 
         private DateTimeFormatFlags formatFlags = DateTimeFormatFlags.NotInitialized;
 
-        private string CultureName => _name ??= _cultureData.CultureName;
+        internal string CultureName => _name ??= _cultureData.CultureName;
 
         private CultureInfo Culture => _cultureInfo ??= CultureInfo.GetCultureInfo(CultureName);