--- a/src/libraries/System.Private.CoreLib/src/System/Globalization/DateTimeParse.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Globalization/DateTimeParse.cs
@@ -3300,7 +3300,7 @@ new DS[] { DS.ERROR,  DS.TX_NNN,  DS.TX_NNN,  DS.TX_NNN,  DS.ERROR,   DS.ERROR, 
             result = -1;
             if (str.GetNext())
             {
-                if (ReferenceEquals(dtfi, DateTimeFormat.InvariantFormatInfo))
+                if (dtfi.CultureName == "")
                 {
                     // Invariant data. Do a fast lookup on the known abbreviated month names.
                     ReadOnlySpan<char> span = str.Value.Slice(str.Index);
@@ -3402,7 +3402,7 @@ new DS[] { DS.ERROR,  DS.TX_NNN,  DS.TX_NNN,  DS.TX_NNN,  DS.ERROR,   DS.ERROR, 
             result = -1;
             if (str.GetNext())
             {
-                if (ReferenceEquals(dtfi, DateTimeFormat.InvariantFormatInfo))
+                if (dtfi.CultureName == "")
                 {
                     // Invariant data. Do a fast lookup on the known month names.
                     ReadOnlySpan<char> span = str.Value.Slice(str.Index);
@@ -3503,7 +3503,7 @@ new DS[] { DS.ERROR,  DS.TX_NNN,  DS.TX_NNN,  DS.TX_NNN,  DS.ERROR,   DS.ERROR, 
             result = -1;
             if (str.GetNext())
             {
-                if (ReferenceEquals(dtfi, DateTimeFormat.InvariantFormatInfo))
+                if (dtfi.CultureName == "")
                 {
                     // Invariant data. Do a fast lookup on the known abbreviated day names.
                     ReadOnlySpan<char> span = str.Value.Slice(str.Index);
@@ -3570,7 +3570,7 @@ new DS[] { DS.ERROR,  DS.TX_NNN,  DS.TX_NNN,  DS.TX_NNN,  DS.ERROR,   DS.ERROR, 
             result = -1;
             if (str.GetNext())
             {
-                if (ReferenceEquals(dtfi, DateTimeFormat.InvariantFormatInfo))
+                if (dtfi.CultureName == "")
                 {
                     // Invariant data. Do a fast lookup on the known day names.
                     ReadOnlySpan<char> span = str.Value.Slice(str.Index);