Speed up the handling of ddd, dddd, MMM, and MMMM parts of a date time format string when using the invariant culture, which is very commonly used in parsing.  Today, when one of these is encountered, the relevant array of comparison strings is retrieved from the DateTimeFormatInfo, and each is compared as a prefix against the current position in the input, using a linguistic ignore-case comparison.  But for the invariant culture, we don't need to consult any arrays, and can do the comparison much more quickly.  These parts dominate the processing of a format like that for RFC1123.

|     Method |         Toolchain |     Mean |    Error |   StdDev | Ratio |
|----------- |------------------ |---------:|---------:|---------:|------:|
| ParseExact | \main\corerun.exe | 997.4 ns | 19.97 ns | 17.70 ns |  1.00 |
| ParseExact |   \pr\corerun.exe | 359.9 ns |  2.99 ns |  2.65 ns |  0.36 |

```C#
const string Format = "ddd, dd MMM yyyy HH':'mm':'ss 'GMT'";

private string _s = new DateTime(1955, 11, 5, 6, 0, 0, DateTimeKind.Utc).ToString(Format, CultureInfo.InvariantCulture);

[Benchmark]
public void ParseExact() => DateTimeOffset.ParseExact(_s, Format, CultureInfo.InvariantCulture, DateTimeStyles.AllowInnerWhite | DateTimeStyles.AssumeUniversal);
```